+++ packages/css-calc/test/wpt/round-function.mjs
@@ -252,12 +252,12 @@ assert.strictEqual(
 
 assert.strictEqual(
 	calc('round(-5, infinity)'),
-	'calc(-1 * 0)',
+	'-0',
 );
 
 assert.strictEqual(
 	calc('round(-5, -infinity)'),
-	'calc(-1 * 0)',
+	'-0',
 );
 
 assert.strictEqual(
@@ -272,12 +272,12 @@ assert.strictEqual(
 
 assert.strictEqual(
 	calc('round(to-zero, -5, infinity)'),
-	'calc(-1 * 0)',
+	'-0',
 );
 
 assert.strictEqual(
 	calc('round(to-zero, -5, -infinity)'),
-	'calc(-1 * 0)',
+	'-0',
 );
 
 // 'up': pos goes to +inf, 0+ goes to 0+, else 0-
@@ -292,13 +292,13 @@ assert.strictEqual(
 );
 
 assert.strictEqual(
-	calc('round(up, -1 * 0, infinity)'),
-	'calc(-1 * 0)',
+	calc('round(up, -0, infinity)'),
+	'-0',
 );
 
 assert.strictEqual(
 	calc('round(up, -1, infinity)'),
-	'calc(-1 * 0)',
+	'-0',
 );
 
 // 'down': neg goes to -inf, -0 goes to -0, else 0+
@@ -308,13 +308,18 @@ assert.strictEqual(
 );
 
 assert.strictEqual(
-	calc('round(down, -1 * 0, infinity)'),
-	'calc(-1 * 0)',
+	calc('round(down, -0, infinity)'),
+	'-0',
 );
 
 assert.strictEqual(
-	calc('round(down, -1 * 0px, calc(1px * infinity))'),
-	'calc(-1 * 0px)',
+	calc('round(down, -0px, calc(1px * infinity))'),
+	'-0px',
+);
+
+assert.strictEqual(
+	calc('round(down, -0%, calc(1% * infinity))'),
+	'-0%',
 );
 
 assert.strictEqual(
@@ -325,4 +330,19 @@ assert.strictEqual(
 assert.strictEqual(
 	calc('round(down, 1, infinity)'),
 	'0',
+);
+
+assert.strictEqual(
+	calc('round(2.3)'),
+	'2',
+);
+
+assert.strictEqual(
+	calc('round(3.7)'),
+	'4',
+);
+
+assert.strictEqual(
+	calc('round(3.7px)'),
+	'round(3.7px)',
 );