+++ packages/css-calc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to CSS Calc
 
+### Unreleased (minor)
+
+- Correctly serialize negative zero `-0` values
+- Add support for `round()` with only a single number argument
+- Add support for `none` in `clamp()`
+
 ### 1.1.7
 
 _February 19, 2024_